+++ Lombiq.Arithmetics.Tests/PositTests/Posit8_0_Tests.cs
@@ -17,7 +17,7 @@ namespace Lombiq.Arithmetics.Tests
 			Assert.AreEqual((int)new Posit8_0(8), 8);
 			Assert.AreEqual((int)new Posit8_0(-16), -16);
 			Assert.AreEqual((int)new Posit8_0(1024), 64);
-						Assert.AreEqual((ulong)new Posit8_0((ulong)63), (ulong)64);
+						Assert.AreEqual((ulong)new Posit8_0((ulong)63), 64U);
 			
 			Assert.AreEqual((int)new Posit8_0(-1024), -64);
